Move syslog4j dependency to the test suite.
