+++ testsuite/integration/pom.xml
@@ -280,6 +280,11 @@
     <dependencies>
         <!-- Dependencies specific to the integration tests. -->
         <dependency>
+            <groupId>org.syslog4j</groupId>
+            <artifactId>syslog4j</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.wildfly</groupId>
             <artifactId>wildfly-webservices-tests-integration</artifactId>
             <version>${project.version}</version>
