+++ testsuite/pom.xml
@@ -154,6 +154,7 @@
 
         <!-- Arquillian dependency versions -->
         <version.arquillian_wildfly>${project.parent.version}</version.arquillian_wildfly>
+        <version.org.syslog4j>0.9.30</version.org.syslog4j>
         <version.saxon>8.7</version.saxon>
 
         <!-- Don't try to deploy the testsuite modules because they don't build jars -->
@@ -180,6 +181,11 @@
                 <version>1.7</version>
                 <scope>system</scope>
                 <systemPath>${java.home}/../lib/tools.jar</systemPath> <!-- since jdk7 also on MAC this is proper path -->
+            </dependency>
+            <dependency>
+                <groupId>org.syslog4j</groupId>
+                <artifactId>syslog4j</artifactId>
+                <version>${version.org.syslog4j}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
